+++ src/stdlib/truncate.rs
@@ -1,10 +1,10 @@
 use crate::compiler::prelude::*;
 
-fn truncate(value: Value, limit: Value, ellipsis: Value) -> Resolved {
+fn truncate(value: Value, limit: Value, suffix: Value) -> Resolved {
     let mut value = value.try_bytes_utf8_lossy()?.into_owned();
     let limit = limit.try_integer()?;
     let limit = if limit < 0 { 0 } else { limit as usize };
-    let ellipsis = ellipsis.try_boolean()?;
+    let suffix = suffix.try_bytes_utf8_lossy()?.to_string();
     let pos = if let Some((pos, chr)) = value.char_indices().take(limit).last() {
         // char_indices gives us the starting position of the character at limit,
         // we want the end position.
@@ -15,9 +15,8 @@ fn truncate(value: Value, limit: Value, ellipsis: Value) -> Resolved {
     };
     if value.len() > pos {
         value.truncate(pos);
-
-        if ellipsis {
-            value.push_str("...");
+        if !suffix.is_empty() {
+            value.push_str(&suffix);
         }
     }
     Ok(value.into())
@@ -44,8 +43,8 @@ impl Function for Truncate {
                 required: true,
             },
             Parameter {
-                keyword: "ellipsis",
-                kind: kind::BOOLEAN,
+                keyword: "suffix",
+                kind: kind::BYTES,
                 required: false,
             },
         ]
@@ -64,8 +63,8 @@ impl Function for Truncate {
                 result: Ok("foo"),
             },
             Example {
-                title: "ellipsis",
-                source: r#"truncate("foo", 2, true)"#,
+                title: "suffix",
+                source: r#"truncate("foo", 2, "...")"#,
                 result: Ok("fo..."),
             },
         ]
@@ -79,12 +78,12 @@ impl Function for Truncate {
     ) -> Compiled {
         let value = arguments.required("value");
         let limit = arguments.required("limit");
-        let ellipsis = arguments.optional("ellipsis").unwrap_or(expr!(false));
+        let suffix = arguments.optional("suffix").unwrap_or(expr!(""));
 
         Ok(TruncateFn {
             value,
             limit,
-            ellipsis,
+            suffix,
         }
         .as_expr())
     }
@@ -94,16 +93,16 @@ impl Function for Truncate {
 struct TruncateFn {
     value: Box<dyn Expression>,
     limit: Box<dyn Expression>,
-    ellipsis: Box<dyn Expression>,
+    suffix: Box<dyn Expression>,
 }
 
 impl FunctionExpression for TruncateFn {
     fn resolve(&self, ctx: &mut Context) -> Resolved {
         let value = self.value.resolve(ctx)?;
         let limit = self.limit.resolve(ctx)?;
-        let ellipsis = self.ellipsis.resolve(ctx)?;
+        let suffix = self.suffix.resolve(ctx)?;
 
-        truncate(value, limit, ellipsis)
+        truncate(value, limit, suffix)
     }
 
     fn type_def(&self, _: &state::TypeState) -> TypeDef {
@@ -126,10 +125,10 @@ mod tests {
              tdef: TypeDef::bytes().infallible(),
          }
 
-        ellipsis {
+        suffix {
             args: func_args![value: "Super",
                              limit: 0,
-                             ellipsis: true
+                             suffix: "..."
             ],
             want: Ok("..."),
             tdef: TypeDef::bytes().infallible(),
@@ -146,7 +145,7 @@ mod tests {
         exact {
             args: func_args![value: "Super",
                              limit: 5,
-                             ellipsis: true
+                             suffix: "..."
             ],
             want: Ok("Super"),
             tdef: TypeDef::bytes().infallible(),
@@ -160,10 +159,10 @@ mod tests {
             tdef: TypeDef::bytes().infallible(),
         }
 
-        big_ellipsis {
+        big_suffix {
             args: func_args![value: "Supercalifragilisticexpialidocious",
                              limit: 5,
-                             ellipsis: true,
+                             suffix: "...",
             ],
             want: Ok("Super..."),
             tdef: TypeDef::bytes().infallible(),
@@ -172,9 +171,18 @@ mod tests {
         unicode {
             args: func_args![value: "♔♕♖♗♘♙♚♛♜♝♞♟",
                              limit: 6,
-                             ellipsis: true
+                             suffix: "..."
             ],
             want: Ok("♔♕♖♗♘♙..."),
+            tdef: TypeDef::bytes().infallible(),
+        }
+
+        alternative_suffix {
+            args: func_args![value: "Super",
+                             limit: 1,
+                             suffix: "[TRUNCATED]"
+            ],
+            want: Ok("S[TRUNCATED]"),
             tdef: TypeDef::bytes().infallible(),
         }
     ];